More minor changes.
